% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{factorize}
\alias{factorize}
\title{Convert any vector into a factor.}
\usage{
factorize(column, levels, labels, ordered = F)
}
\arguments{
\item{column}{A vector to convert into a factor.}

\item{levels}{A vector of levels to accept (optional). In the case of a
labelled object, this input is completely ignored since levels come from
the value labels defined in the labelled object.}

\item{labels}{A vector of labels for the levels (optional). Similarly ignored
in the case of a labelled object.}

\item{ordered}{A boolean indicating whether factor is ordered (optional). If
levels are specified, it uses the given ordering. If not, uses the order it
encounters values in the vector.}
}
\value{
A native R factor, either unordered or ordered.
}
\description{
Convert any vector into a factor.
}
